refactor: removed type prefix in model for repositories
